Making wtlogin.login tlvs

+++ TProtocol/wtlogin/tlvs.cc
@@ -6,7 +6,7 @@
 ld::HexString tlv::MakeTLV(std::string name, std::function<void (ld::HexString &)> writer){
     ld::HexString ret = ""_hex;
     writer(ret);
-    ret.insert<uint16_t>(0, ret.length());
+    ret.insert<uint16_t>(0, (uint16_t)ret.length());
     ret.insert(0, ld::HexString(name));
     return ret;
 }
